feat(ui): enhance UI components with improved default values and type handling across multiple files

+++ lumosai_ui/web-pages/console/history_drawer.rs
@@ -20,7 +20,7 @@ pub fn HistoryDrawer(trigger_id: String, team_id: i32, history: Vec<History>) ->
                             class: "w-full overflow-hidden truncate",
                             a {
                                 class: "block p-2 hover:bg-gray-100 rounded",
-                                href: crate::routes::console::Conversation{team_id, conversation_id: history.id}.to_string(),
+                                href: crate::routes::console::Conversation{team_id, conversation_id: history.id as i64}.to_string(),
                                 {history.summary.clone()}
                             }
                         }
